requirements.txt
streamlit>=1.31.1
sqlalchemy>=2.0.27
python-jose>=3.3.0
passlib>=1.7.4
python-dotenv>=1.0.0
pandas>=2.1.4
matplotlib>=3.8.2
cryptography>=41.0.7
bcrypt>=4.1.2
PyJWT>=2.8.0